# docs/battleships-api-spec.yaml
openapi: 3.0.3
info:
  title: Battleships API
  version: 0.0.1
  description: Web-based system that allows multiple players to play the Battleship game.
servers:
  - url: http://localhost:8888
    description: Local server
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Get all users
      operationId: getUsers
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/orderByParam'
        - $ref: '#/components/parameters/sortDirectionParam'
      responses:
        200:
          description: Object containing all users
          content:
            'application/vnd.siren+json':
              schema:
                allOf:
                  - $ref: '#/components/schemas/Entity'
                  - properties:
                      properties:
                        $ref: '#/components/schemas/GetUsersOutputModel'
        400:
          $ref: '#/components/responses/BadRequest'
        500:
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Users
      summary: Register user
      operationId: register
      requestBody:
        description: User registration information
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserInputModel'
      responses:
        201:
          description: User was created
          content:
            'application/vnd.siren+json':
              schema:
                allOf:
                  - $ref: '#/components/schemas/Entity'
                  - properties:
                      properties:
                        $ref: '#/components/schemas/RegisterUserOutputModel'
        400:
          $ref: '#/components/responses/BadRequest'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /users/refresh-token:
    post:
      tags:
        - Users
      summary: Refreshes a user's token
      operationId: refreshToken
      requestBody:
        description: Refresh token
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenInputModel'
      responses:
        200:
          description: User's token was refreshed
          content:
            'application/vnd.siren+json':
              schema:
                allOf:
                  - $ref: '#/components/schemas/Entity'
                  - properties:
                      properties:
                        $ref: '#/components/schemas/RefreshTokenOutputModel'
        400:
          $ref: '#/components/responses/BadRequest'
        500:
          $ref: '#/components/responses/InternalServerError'
  /users/{username}:
    get:
      tags:
        - Users
      summary: Gets a user by username
      operationId: getUser
      parameters:
        - $ref: '#/components/parameters/usernameParam'
      responses:
        200:
          description: User was found
          content:
            'application/vnd.siren+json':
              schema:
                allOf:
                  - $ref: '#/components/schemas/Entity'
                  - properties:
                      properties:
                        $ref: '#/components/schemas/GetUserOutputModel'
        400:
          $ref: '#/components/responses/BadRequest'
        500:
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    RegisterUserOutputModel:
      type: object
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzUxMiJ9.eyJ1c2VybmFtZSI6Imd1ZXN0MSIsImlhdCI6MTY2NjM4NzEzNywiZXhwIjoxNjY2MzkwNzM3fQ.Dk3vJVZfeldVPcLSsmAduu52AOS7EyvCkzeDJMcPPhHS31hI8N0mXQeqJUmY3V_o26msseYmzqXf4GvPf33tNA
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzUxMiJ9.eyJ1c2VybmFtZSI6Imd1ZXN0MSIsImlhdCI6MTY2NjM4NzEzNywiZXhwIjoxNjY2NDczNTM3fQ.n38mWAZVW6sfCVkvxfKrwerYKhTXdkAxGVPUikiGwldQPWwYNZ0oMRbkWPQIdi0cP1NDgklS8Da6RchdgdOUaw
    RegisterUserInputModel:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          example: guest1
        email:
          type: string
          example: guest1@mail.com
        password:
          type: string
          example: guest
    RefreshTokenOutputModel:
      type: object
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzUxMiJ9.eyJ1c2VybmFtZSI6Imd1ZXN0MSIsImlhdCI6MTY2NjM4OTc0OCwiZXhwIjoxNjY2MzkzMzQ4fQ.Zh6mjDKVbBH239u3l0AAGnBnQdPrvDsf6qubJsILDquDh8Wec3YAMKIB2uzcNElBVEhKN8QAMhldROhtYGYncg
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzUxMiJ9.eyJ1c2VybmFtZSI6Imd1ZXN0MSIsImlhdCI6MTY2NjM4OTc0OCwiZXhwIjoxNjY2NDc2MTQ4fQ.hjkkN1ORnLC0Qmp0xEpElhGMwit3CnW7dDBzibMUMd3qJVNkoJVrSp9G4AwizAMeF1v28s3nD5OVERzlNDvDLQ
    RefreshTokenInputModel:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzUxMiJ9.eyJ1c2VybmFtZSI6Imd1ZXN0MSIsImlhdCI6MTY2NjM4OTc0OCwiZXhwIjoxNjY2NDc2MTQ4fQ.hjkkN1ORnLC0Qmp0xEpElhGMwit3CnW7dDBzibMUMd3qJVNkoJVrSp9G4AwizAMeF1v28s3nD5OVERzlNDvDLQ
...
    GetUsersOutputModel:
      type: object
      properties:
        totalCount:
          type: integer
          example: 10
    GetUserOutputModel:
      type: object
      properties:
        username:
          type: string
          example: guest1
        email:
          type: string
          example: guest1@guest.org
        points:
          type: integer
          example: 100
...
    Entity:
      type: object
      properties:
        class:
          type: array
          items:
            type: string
        properties:
          type: object
          additionalProperties: false
        entities:
          type: array
          items:
            $ref: '#/components/schemas/SubEntity'
        actions:
          type: array
          items:
            $ref: '#/components/schemas/Action'
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    SubEntity:
      type: object
      required:
        - rel
      properties:
        class:
          type: array
          items:
            type: string
        rel:
          type: array
          items:
            type: string
        href:
          type: string
        type:
          type: string
        title:
          type: string
        properties:
          type: object
          additionalProperties: false
        entities:
          type: array
          items:
            $ref: '#/components/schemas/SubEntity'
        actions:
          type: array
          items:
            $ref: '#/components/schemas/Action'
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    Action:
      type: object
      required:
        - name
        - href
      properties:
        name:
          type: string
        title:
          type: string
        method:
          type: string
        href:
          type: string
        type:
          type: string
        fields:
          type: array
          items:
            $ref: '#/components/schemas/Field'
    Field:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        type:
          type: string
        value:
          type: string
        title:
          type: string
    Link:
      type: object
      required:
        - rel
        - href
      properties:
        rel:
          type: array
          items:
            type: string
        href:
          type: string
        type:
          type: string
        title:
          type: string
    ErrorModel:
      type: object
      properties:
        timestamp:
          type: string
        status:
          type: integer
        error:
          type: string
        path:
          type: string
  parameters:
    offsetParam:
      name: offset
      in: query
      description: The offset to start from
      required: true
      schema:
        type: integer
        default: 0
    limitParam:
      name: limit
      in: query
      description: The number of items to return
      required: true
      schema:
        type: integer
        default: 10
    orderByParam:
      name: orderBy
      in: query
      description: The field to order by
      required: false
      schema:
        type: string
        enum:
          - points
          - number_of_games_played
        default: points
    sortDirectionParam:
      name: sortDirection
      in: query
      description: The direction to sort by
      required: false
      schema:
        type: string
        enum:
          - ASC
          - DESC
        default: ASC
...
    usernameParam:
      name: username
      in: path
      description: The username of the user
      required: true
      schema:
        type: string
        example: 'bob'
  responses:
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT